requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
Pillow==10.2.0
python-escpos==2.2.0 